// packages/stdlib/src/constants/action-failure-reason.d.ts
/**
 * Enum for common action failure reasons in the IF system.
 *
 * These are used by actions to indicate why an action failed,
 * allowing the text service to map them to appropriate messages
 * in the current language.
 */
export declare enum ActionFailureReason {
    NOT_VISIBLE = "NOT_VISIBLE",
    NOT_REACHABLE = "NOT_REACHABLE",
    NOT_IN_SCOPE = "NOT_IN_SCOPE",
    FIXED_IN_PLACE = "FIXED_IN_PLACE",
    ALREADY_OPEN = "ALREADY_OPEN",
    ALREADY_CLOSED = "ALREADY_CLOSED",
    NOT_OPENABLE = "NOT_OPENABLE",
    LOCKED = "LOCKED",
    NOT_LOCKABLE = "NOT_LOCKABLE",
    ALREADY_LOCKED = "ALREADY_LOCKED",
    ALREADY_UNLOCKED = "ALREADY_UNLOCKED",
    CONTAINER_FULL = "CONTAINER_FULL",
    CONTAINER_CLOSED = "CONTAINER_CLOSED",
    NOT_A_CONTAINER = "NOT_A_CONTAINER",
    NOT_A_SUPPORTER = "NOT_A_SUPPORTER",
    ALREADY_IN_CONTAINER = "ALREADY_IN_CONTAINER",
    NOT_WEARABLE = "NOT_WEARABLE",
    ALREADY_WEARING = "ALREADY_WEARING",
    NOT_WEARING = "NOT_WEARING",
    WORN_BY_OTHER = "WORN_BY_OTHER",
    TOO_HEAVY = "TOO_HEAVY",
    CARRYING_TOO_MUCH = "CARRYING_TOO_MUCH",
    WRONG_KEY = "WRONG_KEY",
    NO_KEY_SPECIFIED = "NO_KEY_SPECIFIED",
    NOT_A_KEY = "NOT_A_KEY",
    ALREADY_ON = "ALREADY_ON",
    ALREADY_OFF = "ALREADY_OFF",
    NOT_SWITCHABLE = "NOT_SWITCHABLE",
    NO_EXIT_THAT_WAY = "NO_EXIT_THAT_WAY",
    CANT_GO_THAT_WAY = "CANT_GO_THAT_WAY",
    DOOR_CLOSED = "DOOR_CLOSED",
    DOOR_LOCKED = "DOOR_LOCKED",
    CANT_TALK_TO_THAT = "CANT_TALK_TO_THAT",
    NO_RESPONSE = "NO_RESPONSE",
    NOT_A_PERSON = "NOT_A_PERSON",
    CANT_DO_THAT = "CANT_DO_THAT",
    NOT_IMPLEMENTED = "NOT_IMPLEMENTED",
    INVALID_TARGET = "INVALID_TARGET",
    AMBIGUOUS_TARGET = "AMBIGUOUS_TARGET",
    NOTHING_HAPPENS = "NOTHING_HAPPENS",
    ACTOR_CANT_SEE = "ACTOR_CANT_SEE",
    ACTOR_CANT_REACH = "ACTOR_CANT_REACH",
    ACTOR_BUSY = "ACTOR_BUSY",
    NOT_EDIBLE = "NOT_EDIBLE",
    NOT_READABLE = "NOT_READABLE",
    NOTHING_WRITTEN = "NOTHING_WRITTEN",
    WONT_ACCEPT = "WONT_ACCEPT",
    CANT_GIVE_TO_SELF = "CANT_GIVE_TO_SELF",
    CANT_USE_THAT = "CANT_USE_THAT",
    CANT_USE_TOGETHER = "CANT_USE_TOGETHER",
    NOTHING_TO_USE_WITH = "NOTHING_TO_USE_WITH"
}
/**
 * Type guard to check if a value is an ActionFailureReason
 */
export declare function isActionFailureReason(value: any): value is ActionFailureReason;
//# sourceMappingURL=action-failure-reason.d.ts.map